# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Memora API (mock)
  version: "1.0.0"
  description: |
    API mock do Memora — Express 5, estado em memória (reinicia com o processo, sem banco de
    dados). Reproduz as regras de negócio do Escopo Consolidado v2 para servir de backend ao
    protótipo navegável; não é a API de produção.

    Todas as rotas de negócio vivem sob `/v1`. A autenticação é por sessão simples (token opaco
    em `Authorization: Bearer <token>`), sem JWT nem refresh — ver
    [ADR-0008](./adr/0008-autenticacao-por-sessao-simples.md).

    Cada área do produto (aluno, responsável, B2B, back-office, parceiro) faz login com sua
    própria rota `POST .../auth/login` e recebe um token cujo `role` restringe o que pode
    acessar (`requireRole` em `backend/src/lib/auth.js`).
  contact:
    name: Memora — engenharia
  license:
    name: Uso interno
servers:
  - url: http://localhost:3180/v1
    description: Desenvolvimento local (porta configurável via `PORT`)
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: >
        Token de sessão retornado por qualquer rota `auth/login`. Sessões expiram em 30 dias
        sem uso (deslizante); usar em `me`, `parental`, `b2b`, `admin` e `partner` é mutuamente
        exclusivo por `role`.
...
  responses:
    Unauthorized:
      description: Sessão ausente, inválida ou expirada.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { code: unauthorized, message: "Sessão inválida ou expirada." }
...
    Validation:
      description: Corpo da requisição inválido.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string, description: "Slug estável da causa, ex.: validation, not_found, guardian_required." }
        message: { type: string, description: "Mensagem em PT-BR, pronta para exibição." }
      additionalProperties: true
      description: Formato uniforme de erro em toda a API (ver `fail()` em `backend/src/lib/helpers.js`).
    User:
      type: object
      description: Perfil público do aluno logado (`publicUser()` em `backend/src/lib/helpers.js`).
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string }
        phone: { type: string }
        birthDate: { type: string, format: date }
        isMinor: { type: boolean }
        guardianId: { type: string, nullable: true }
        plan: { type: string, enum: [free, premium] }
        locale: { type: string, example: pt-BR }
        timezone: { type: string, example: America/Sao_Paulo }
        verified:
          type: object
          properties: { email: { type: boolean }, phone: { type: boolean } }
        canEarn: { type: boolean, description: "true somente com e-mail e telefone confirmados." }
        ageVerification:
          type: object
          properties:
            status: { type: string, enum: [pendente, aguardando_responsavel, aprovado] }
            method: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        deactivatedAt: { type: string, format: date-time, nullable: true }
        publicProfile: { type: object, additionalProperties: true }
        settings: { type: object, additionalProperties: true }
        guardianPermissions: { type: object, nullable: true, additionalProperties: true }
        streak: { type: integer }
        studyDays: { type: integer }
        balance: { type: integer, description: "Saldo de Neurônios." }
        onboardingDone: { type: boolean }
...
    Objective:
      type: object
      properties:
        id: { type: string }
        userId: { type: string }
        name: { type: string }
        date: { type: string, format: date }
        subjects: { type: array, items: { type: string } }
        outcome: { type: string, nullable: true, enum: [aprovado, reprovado, aguardando, nao_fiz, null] }
        outcomeRewarded: { type: boolean }
        daysLeft: { type: integer }
        past: { type: boolean }
        createdAt: { type: string, format: date-time }
...
paths:
...
  /auth/login:
    post:
      tags: [Autenticação]
      summary: Login do aluno (e-mail ou atalho de demonstração)
      security: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                provider: { type: string, enum: [email, google, microsoft], default: email }
                email: { type: string }
                password: { type: string }
                device: { type: string }
      responses:
        "200":
          description: Sessão criada. Qualquer credencial preenchida funciona no protótipo.
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  sessionId: { type: string }
                  user: { $ref: '#/components/schemas/User' }
        "422": { $ref: '#/components/responses/Validation' }
  /auth/register:
    post:
      tags: [Autenticação]
      summary: Cadastro (telefone e data de nascimento obrigatórios)
      security: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [name, email, phone, birthDate, student]
              properties:
                name: { type: string }
                email: { type: string }
                phone: { type: string, description: "10 a 11 dígitos com DDD." }
                birthDate: { type: string, format: date }
                student: { type: boolean }
                origin: { type: string, nullable: true }
                captchaToken: { type: string, nullable: true }
      responses:
        "201":
          description: Conta criada; próximo passo é confirmar e-mail/telefone.
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  user: { $ref: '#/components/schemas/User' }
                  age: { type: integer, nullable: true }
                  requiresGuardianConsent: { type: boolean }
                  nextStep: { type: string, example: confirm }
        "422": { $ref: '#/components/responses/Validation' }
        "429":
          description: "Mais de 5 contas pelo mesmo IP em 24 h sem `captchaToken`."
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  /auth/confirm:
    post:
      tags: [Autenticação]
      summary: "Confirma e-mail ou telefone (código fixo no protótipo: 123456)"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [channel, code]
              properties:
                channel: { type: string, enum: [email, phone] }
                code: { type: string }
      responses:
        "200":
          description: Canal confirmado.
          content:
            application/json:
              schema:
                type: object
                properties:
                  verified: { type: object, properties: { email: { type: boolean }, phone: { type: boolean } } }
                  user: { $ref: '#/components/schemas/User' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "422": { $ref: '#/components/responses/Validation' }
  /auth/resend:
    post:
      tags: [Autenticação]
      summary: Reenvia o código de confirmação
      requestBody:
        content:
          application/json:
            schema: { type: object, properties: { channel: { type: string, enum: [email, phone] } } }
      responses:
        "200": { description: Reenviado (sempre no mock)., content: { application/json: { schema: { type: object, properties: { sent: { type: boolean }, channel: { type: string } } } } } }
  /auth/age-verification:
    post:
      tags: [Autenticação]
      summary: Verificação de idade por terceiro homologado
      description: >
        Menores de 16 anos exigem dados do responsável (`guardian.name`) antes de aprovar.
        Retenção: apenas resultado, método, data, hash e verificador — documento é apagado em 24 h.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                method: { type: string }
                guardian: { type: object, properties: { name: { type: string }, email: { type: string }, phone: { type: string } } }
      responses:
        "200":
          description: "Aprovado, ou `aguardando_responsavel` se faltar o responsável."
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, enum: [aprovado, aguardando_responsavel] }
                  ageVerification: { type: object, additionalProperties: true }
                  isMinor: { type: boolean }
                  user: { $ref: '#/components/schemas/User' }
  /auth/onboarding:
    post:
      tags: [Autenticação]
      summary: Onboarding — contexto, matérias, tempo disponível e meta sugerida
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                context: { type: string }
                subjects: { type: array, items: { type: string } }
                minutesPerDay: { type: integer, default: 10 }
                objective: { type: object, properties: { name: { type: string }, date: { type: string, format: date } } }
      responses:
        "200":
          description: Meta diária calculada e objetivo opcional criado.
          content:
            application/json:
              schema:
                type: object
                properties:
                  goal: { type: integer }
                  objective: { $ref: '#/components/schemas/Objective' }
                  planOfDay: { type: object, properties: { minutes: { type: integer }, cards: { type: integer } } }
  /auth/password/recover:
    post:
      tags: [Autenticação]
      summary: Recuperação de senha por e-mail
      security: []
      requestBody:
        content: { application/json: { schema: { type: object, properties: { email: { type: string } } } } }
      responses:
        "200": { description: "Sempre responde `sent: true` (não revela se a conta existe)." }
        "422": { $ref: '#/components/responses/Validation' }
...